refactor(@xen-orchestra/backups): split Backup Job in XapiVmBackupJob and MetadatasBackupJob
